Consolidate frontend API layer
