--- frontend/src/api/index.js
@@ -1,41 +1,0 @@
-import axios from 'axios';
-
-// Create an axios instance with default settings
-const api = axios.create({
-  baseURL: process.env.REACT_APP_API_URL || '/api',
-  headers: {
-    'Content-Type': 'application/json'
-  }
-});
-
-// Add a request interceptor to add Authorization header
-api.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem('token');
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => Promise.reject(error)
-);
-
-// Add a response interceptor to handle errors
-api.interceptors.response.use(
-  (response) => response,
-  (error) => {
-    // Handle session expiration
-    if (error.response && error.response.status === 401) {
-      if (window.location.pathname !== '/login' && 
-          window.location.pathname !== '/register' && 
-          !window.location.pathname.includes('/reset-password')) {
-        localStorage.removeItem('token');
-        localStorage.removeItem('user');
-        window.location.href = '/login?session=expired';
-      }
-    }
-    return Promise.reject(error);
-  }
-);
-
-export default api;
