+++ frontend/src/api/notifications.js
@@ -1,4 +1,4 @@
-import api from './index';
+import api from '../services/api/client';
 
 const NotificationsAPI = {
   // Get all notifications
