+++ frontend/src/api/referrals.js
@@ -1,4 +1,4 @@
-import api from './index';
+import api from '../services/api/client';
 
 const ReferralsAPI = {
   // Create a new referral
